+++ data-manager/api/src/main/java/org/orbisgis/orbisdata/datamanager/api/dataset/IDataSet.java
@@ -42,10 +42,13 @@ import org.orbisgis.commons.annotations.Nullable;
 /**
  * Raw collection of data, no matter its structure.
  *
+ * @param <T> The type of elements returned by the iterator.
+ * @param <U> The type of elements streamed.
+ *
  * @author Erwan Bocher (CNRS)
  * @author Sylvain PALOMINOS (UBS 2018-2019)
  */
-public interface IDataSet<T> extends Iterable<T> {
+public interface IDataSet<T, U> extends Iterable<T> {
 
     /**
      * Get the location of {@link IDataSet}.
@@ -110,7 +113,7 @@ public interface IDataSet<T> extends Iterable<T> {
      * @param filter Filtes to apply.
      * @return A filtered {@link IDataSet}
      */
-    IDataSet<?> filter(String filter);
+    IDataSet<?, ?> filter(String filter);
 
     @Override
     @NotNull