+++ data-manager/jdbc/src/main/java/org/orbisgis/orbisdata/datamanager/jdbc/h2gis/H2gisSpatialTable.java
@@ -110,7 +110,7 @@ public class H2gisSpatialTable extends JdbcSpatialTable {
     }
 
     @Override
-    public ISpatialTable<Object> reproject(int srid) {
+    public ISpatialTable reproject(int srid) {
         try {
             ResultSetMetaData meta = getMetaData();
             if(meta == null){